requirements.txt
discord.py==1.5.1
Pillow ==8.0.1
requests==2.24.0
PyNaCl==1.3.0
dnspython==1.16.0